fix: rebase 버그 수정
