+++ client/src/components/item/main-item-wrapper.tsx
@@ -1,7 +1,8 @@
 import React, { FC } from 'react';
 import styled from 'lib/woowahan-components';
 import { IItem } from 'types/item';
-import ItemList from 'components/item/item-list';
+import { Loader } from 'components';
+import DisplayItemList from './display-item-list';
 
 interface ItemListProps {
   popularItems: IItem[] | null;
@@ -10,43 +11,19 @@ interface ItemListProps {
   loading: boolean;
 }
 
-const Div = styled.div`
-  margin-top: 90px;
-  ${props => props.theme?.mobile} {
-    margin-top: 50px;
-  }
-
-  ${props => props.theme?.tablet} {
-    margin-top: 70px;
-  }
+const Wrapper = styled.div`
+  margin-top: 60px;
 `;
 
-const SmallTitle = styled.div`
-  font-family: ${props => props.theme?.fontHanna};
-  font-size: 36px;
-  ${props => props.theme?.mobile} {
-    font-size: 20px;
-    margin-left: 50px;
-  }
-
-  ${props => props.theme?.tablet} {
-    font-size: 28px;
-    margin-left: 50px;
-  }
-`;
-
-const MainItemWrapper: FC<ItemListProps> = ({ popularItems, newItems, recommendItems, loading }) => {
+const MainItems: FC<ItemListProps> = ({ popularItems, newItems, recommendItems, loading }) => {
   return (
-    <Div>
-      {loading && <div>로딩중</div>}
-      <SmallTitle>잘나가요</SmallTitle>
-      <ItemList items={popularItems} isLoading={loading} />
-      <SmallTitle>새로 나왔어요</SmallTitle>
-      <ItemList items={newItems} isLoading={loading} />
-      <SmallTitle>추천 드려요</SmallTitle>
-      <ItemList items={recommendItems} isLoading={loading} />
-    </Div>
+    <Wrapper>
+      {loading && <Loader size="25px" color="brown" />}
+      <DisplayItemList title="잘나가요" items={popularItems} />
+      <DisplayItemList title="새로 나왔어요" items={newItems} />
+      <DisplayItemList title="추천드려요" items={recommendItems} />
+    </Wrapper>
   );
 };
 
-export default MainItemWrapper;
+export default MainItems;
